PD-1: Apply code review feedback and extract methods into separate classes

+++ src/main/java/pd1/Player.java
@@ -8,11 +8,9 @@ public class Player {
         this.name = name;
         this.scores = scores;
     }
-
     public String getName() {
         return name;
     }
-
     public int getTotalScore() {
         int totalScore = 0;
         for (int score : scores) {
@@ -20,23 +18,19 @@ public class Player {
         }
         return totalScore;
     }
-
     public double getAverageScore() {
-        double averageCalculation = (double) getTotalScore() / scores.length;
+        double averageCalculation = (double) getTotalScore() / (double) scores.length;
         return averageCalculation;
     }
-
     public int getMinScore() {
         int minScore = scores[0];
         for (int score : scores) {
             if (score < minScore) {
                 minScore = score;
-
             }
         }
         return minScore;
     }
-
     public int getMaxScore() {
         int maxScore = scores[0];
         for (int score : scores) {
